Move twitter code into friend graphs module
